change Percentage Value from decimal to integer in routing processes
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C43219

+++ TOneV2/Code/TOneV2/TOne.WhS.DBSync.Business/Migrators/Rules/SpecialRequestRuleMigrator.cs
@@ -190,7 +190,7 @@ namespace TOne.WhS.DBSync.Business
             {
                 ForceOption = option.ForcedOption,
                 NumberOfTries = option.NumberOfTries,
-                Percentage = option.Percentage == 0 ? (decimal?)null : option.Percentage,
+                Percentage = option.Percentage == 0 ? (int?)null : option.Percentage,
                 Position = ++position,
                 SupplierId = supplier.CarrierAccountId
             };
