+++ TOneV2/Code/TOneV2/TOne.WhS.Routing.Entities/Routes/RPRouteOption.cs
@@ -18,7 +18,7 @@ namespace TOne.WhS.Routing.Entities
 
         public Decimal SupplierRate { get; set; }
 
-        public Decimal? Percentage { get; set; }
+        public int? Percentage { get; set; }
 
         public decimal OptionWeight { get; set; }
