+++ TOneV2/Code/TOneV2/TOne.WhS.Routing.Entities/SpecialRequestRouteRule/SpecialRequestRouteOptionSettings.cs
@@ -16,6 +16,6 @@ namespace TOne.WhS.Routing.Entities
 
         public int NumberOfTries { get; set; }
 
-        public Decimal? Percentage { get; set; }
+        public int? Percentage { get; set; }
     }
 }
